backend/requirements.txt
fastapi[standard]
pandas
numpy
scikit-learn
plotly
langchain
langchain-google-genai
langchain-experimental
python-multipart
openpyxl
lxml
pydantic
python-dotenv
tabulate
google-generativeai
pymongo
dnspython
seaborn
